# openapi/site.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OneMail Site API",
    "version": "1.0.0"
  },
...
  "paths": {
    "/auth/direct/login": {
      "post": {
        "operationId": "SiteAuth_directLogin",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDirectLoginResult"
                }
              }
            }
          },
          "400": {
            "description": "RFC 7807 bad request response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Access is forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDirectLoginError"
                }
              }
            }
          }
        },
        "tags": [
          "Site Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteDirectLoginInput"
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "SiteAuth_register",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteRegisterResult"
                }
              }
            }
          },
          "409": {
            "description": "RFC 7807 conflict response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "RFC 7807 validation response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteRegisterInput"
              }
            }
          }
        },
        "security": [
          {}
        ]
      }
    },
    "/contacts": {
      "post": {
        "operationId": "SiteContacts_create",
        "description": "Create a contact from the site UI",
        "parameters": [],
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteContactResource"
                }
              }
            }
          },
          "409": {
            "description": "RFC 7807 conflict response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "RFC 7807 validation response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Contacts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCreateContactInput"
              }
            }
          }
        }
      },
      "get": {
        "operationId": "SiteContacts_list",
        "description": "List contacts for the site UI",
        "parameters": [
          {
            "$ref": "#/components/parameters/PageQuery.page"
          },
          {
            "$ref": "#/components/parameters/PageQuery.pageSize"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by status",
            "schema": {
              "$ref": "#/components/schemas/SiteContactStatus"
            },
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "items",
                    "page",
                    "pageSize",
                    "totalItems",
                    "totalPages"
                  ],
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/SiteContactResource"
                      },
                      "description": "List of items"
                    },
                    "page": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Page number (1-based)"
                    },
                    "pageSize": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Page size"
                    },
                    "totalItems": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Total number of elements"
                    },
                    "totalPages": {
                      "type": "integer",
                      "format": "int32",
                      "description": "Total number of pages"
                    }
                  },
                  "description": "Paginated response"
                }
              }
            }
          },
          "400": {
            "description": "RFC 7807 bad request response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "RFC 7807 validation response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Contacts"
        ]
      }
    },
    "/contacts/{id}": {
      "get": {
        "operationId": "SiteContacts_get",
        "description": "Get a contact by ID for the site UI",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteContactResource"
                }
              }
            }
          },
          "400": {
            "description": "RFC 7807 bad request response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "RFC 7807 not found response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Contacts"
        ]
      },
      "put": {
        "operationId": "SiteContacts_update",
        "description": "Update a contact from the site UI",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteContactResource"
                }
              }
            }
          },
          "400": {
            "description": "RFC 7807 bad request response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "RFC 7807 not found response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "RFC 7807 conflict response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "RFC 7807 validation response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Contacts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteUpdateContactInput"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "SiteContacts_delete",
        "description": "Delete a contact from the site UI",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityId"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful."
          },
          "400": {
            "description": "RFC 7807 bad request response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "RFC 7807 not found response",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "tags": [
          "Site Contacts"
        ]
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "parameters": {
      "PageQuery.page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Page number (1-based)",
        "schema": {
          "type": "integer",
          "format": "int32",
          "default": 1
        },
        "explode": false
      },
      "PageQuery.pageSize": {
        "name": "pageSize",
        "in": "query",
        "required": false,
        "description": "Page size",
        "schema": {
          "type": "integer",
          "format": "int32",
          "default": 25
        },
        "explode": false
      }
    },
    "schemas": {
      "EmailAddress": {
        "type": "string",
        "format": "email"
      },
      "EntityId": {
        "type": "string",
        "pattern": "^[0-9]+$"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code"
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence"
          },
          "instance": {
            "type": "string",
            "description": "A URI reference that identifies the specific occurrence"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "description": "Validation errors grouped by field"
          },
          "form": {
            "type": "string",
            "description": "Form-level validation error"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Field validation errors"
          }
        },
        "description": "RFC 7807 Problem Details"
      },
      "SiteContactResource": {
        "type": "object",
        "required": [
          "id",
          "email",
          "status",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EntityId"
              }
            ],
            "description": "Unique identifier"
          },
          "email": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmailAddress"
              }
            ],
            "description": "Email address"
          },
          "firstName": {
            "type": "string",
            "nullable": true,
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "description": "Last name"
          },
          "timeZone": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeZoneName"
              }
            ],
            "nullable": true,
            "description": "IANA time zone identifier"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true,
            "description": "Custom fields as key-value pairs"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SiteContactStatus"
              }
            ],
            "description": "Current status"
          },
          "createdAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ],
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Timestamp"
              }
            ],
            "description": "Last update timestamp"
          }
        },
        "description": "Contact resource used by the site UI"
      },
      "SiteContactStatus": {
        "type": "string",
        "enum": [
          "active",
          "unsubscribed"
        ],
        "description": "Contact status for site UI"
      },
      "SiteCreateContactInput": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "allOf": [
              {
                "$ref": "#/components/schemas/EmailAddress"
              }
            ],
            "description": "Email address"
          },
          "firstName": {
            "type": "string",
            "nullable": true,
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "description": "Last name"
          },
          "timeZone": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeZoneName"
              }
            ],
            "nullable": true,
            "description": "IANA time zone identifier"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true,
            "description": "Custom fields as key-value pairs"
          }
        },
        "description": "Site request body for creating a contact"
      },
      "SiteDirectLoginError": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "SiteDirectLoginInput": {
        "type": "object",
        "required": [
          "user",
          "passwd"
        ],
        "properties": {
          "user": {
            "type": "string"
          },
          "passwd": {
            "type": "string"
          }
        }
      },
      "SiteDirectLoginResult": {
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "picture": {
            "type": "string"
          },
          "aud": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "attrs": {
            "type": "object",
            "additionalProperties": {}
          },
          "role": {
            "type": "string"
          }
        }
      },
      "SiteRegisterInput": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "$ref": "#/components/schemas/EmailAddress"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "SiteRegisterResult": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "createdAt"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/EntityId"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "$ref": "#/components/schemas/EmailAddress"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Timestamp"
          }
        }
      },
      "SiteUpdateContactInput": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "nullable": true,
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "nullable": true,
            "description": "Last name"
          },
          "timeZone": {
            "type": "string",
            "allOf": [
              {
                "$ref": "#/components/schemas/TimeZoneName"
              }
            ],
            "nullable": true,
            "description": "IANA time zone identifier"
          },
          "customFields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true,
            "description": "Custom fields as key-value pairs"
          }
        },
        "description": "Site request body for updating a contact"
      },
      "TimeZoneName": {
        "type": "string"
      },
      "Timestamp": {
        "type": "string",
        "format": "date-time"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development site API",
      "variables": {}
    }
  ]
}